### 2022-06-09 #4463 @editorialbot
[PRE REVIEW]: stacks: Stacked Ensemble Modeling with Tidy Data Principles
**Submitting author:** @simonpcouch (<a href="http://orcid.org/0000-0001-5676-5107">Simon Couch</a>)
**Repository:** https://github.com/tidymodels/stacks
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.3
