### 2022-06-13 #4463 @mllg
I'm author of a similar package (mlr3pipelines, https://mlr3book.mlr-org.com/pipelines.html#pipe-model-ensembles-stacking). In order to not give the appearance of a conflict of interests, I would withdraw unless I'm really needed.
