### 2022-06-15 tidymodels/stacks#135 @rcannood
[JOSS] Comments
Hi all!

I'm currently reviewing the JOSS submission for stacks (https://github.com/openjournals/joss-reviews/issues/4471). I think the paper is *very* well written and is worthy of publication at JOSS.

Below is an overview of all my relatively minor comments. I'm happy to discuss them in this thread or in a separate issue, if need be.

Robrecht

---

- [x] State of the field: I think it would be worthwhile to include mlr3 in the comparison to other packages.
- [x] State of the field: The 'Statement of need' and 'Comparison to other packages' sections only discuss model stacking / ensemble modelling in the context of R. Please also discuss the state of the field outside the R ecosystem?
- [x] References: It probably makes sense to cite the academic paper for superlearner (https://www.degruyter.com/document/doi/10.2202/1544-6115.1309/html?) and H2O (https://www.automl.org/wp-content/uploads/2020/07/AutoML_2020_paper_61.pdf?).
