### 2022-06-16 tidymodels/stacks#135 @simonpcouch
@rcannood, thank you for reviewing! I understand + support your comments. I'll incorporate your suggestions into the paper draft early next week.
