### 2022-06-21 a2576e5
address JOSS review comments (#135)
* include mlr3 in package comparisons
* reference non-R implementations
* refine citations
* change relevant section header accordingly
