### 2022-06-21 tidymodels/stacks#135 @simonpcouch
Thanks again! Just implemented your suggestions—will update the paper draft on the review issue.
