### 2022-06-21 #4471 @simonpcouch
Updating with changes in response to https://github.com/tidymodels/stacks/issues/135!
