### 2022-06-21 #4471 @simonpcouch
re: the reported missing DOI, the suggested DOI for Keras [doesn't seem to point to the deep learning framework](https://referenceworks.brillonline.com/entries/der-neue-pauly/*-e612900), and the bib entry for this paper uses the [requested citation](https://keras.io/getting_started/faq/#how-should-i-cite-keras) in Keras docs.
