+++ dev/pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.14.2
-pkgdown: 2.0.3
+pkgdown: 2.0.4
 pkgdown_sha: ~
 articles:
   basics: basics.html
   classification: classification.html
-last_built: 2022-06-09T17:01Z
+last_built: 2022-06-21T18:20Z
 urls:
   reference: https://stacks.tidymodels.org/reference
   article: https://stacks.tidymodels.org/articles
