### 2022-06-22 bbf7c75
subset hard class predictions by level rather than index
fixes bug where stacks will return incorrect predictions if an elastic net meta-learner is used, the `type` argument to `predict` is set to `"class"`, and the outcome levels differ from alphabetical order.
