+++ dev/pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   basics: basics.html
   classification: classification.html
-last_built: 2022-06-21T18:20Z
+last_built: 2022-06-22T15:25Z
 urls:
   reference: https://stacks.tidymodels.org/reference
   article: https://stacks.tidymodels.org/articles
