### 2022-06-30 tidymodels/stacks#135 @rcannood
Looks good! Thanks for taking care of this so quickly :)
