### 2022-06-30 #4471 @rcannood
Hi @simonpcouch @osorensen!

As mentioned earlier, I think the paper is very well written and is definitely worthy of publication at JOSS.

I had a few minor comments (tidymodels/stacks#135), but @simonpcouch was able to resolve them quickly. I have no further remarks.

Looking forward to seeing @mcavs' comments :)
