### 2022-07-02 tidymodels/stacks#137 @mcavs
JOSS Review for stacks: Stacked Ensemble Modeling with Tidy Data Principles #4471
Dear @simonpcouch,

Your paper is submitted to JOSS ([https://github.com/openjournals/joss-reviews/issues/4471](https://github.com/openjournals/joss-reviews/issues/4471)), is well-written and pretty understandable. But I think it needs some minor additions:

* The statement of need and the mechanism of the package look enough convincing, but there is no link or example of how it can be used. To link or cite [the vignettes](https://stacks.tidymodels.org/articles/basics.html) of the {stacks} may be useful for the readers.

* It is indicated the main superiority of the {stacks} is its tidy mechanism that utilizes the consistent and unified interface of the {tidymodels} packages to implement a generalized and concise grammar for model stacking. In the section "Comparison to Other Softwares", some specific comparisons between its alternatives may be given besides general comments.

Thanks,
Mustafa
