### 2022-07-02 #4471 @simonpcouch
Absolutely—will address https://github.com/tidymodels/stacks/issues/137 first thing next week! Thanks all for your work.
