### 2022-07-05 6ebcf15
address JOSS review comments (#137)
* link out to package documentation/vignettes
* make specific comparisons to other R ensembling implementations
