### 2022-07-05 c228047
install tidymodels for website vignette builds
