+++ DESCRIPTION
@@ -58,7 +58,6 @@ Suggests:
     SuperLearner
 Config/Needs/website: 
     tidyverse/tidytemplate 
-    tidymodels/tidymodels
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
