### 2022-07-05 #4471 @simonpcouch
Thank you! Of course.

The tag for the archived release is [v0.2.4](https://github.com/tidymodels/stacks/releases/tag/v0.2.4). The DOI is [10.5281/zenodo.6800026](https://doi.org/10.5281/zenodo.6800026).
