### 2022-07-05 tidymodels/stacks#138 @osorensen
Editorial comments on JOSS submission
Here are some editorial comments on the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4471).

In general, the paper reads very well. My main point is that all software packages mentioned should be properly cited, at least the first time they're mentioned, even when it is a reference to CRAN. Here are the ones I noticed, but I might have missed some:

- [ ] Lines 44-45, add the references for {parsnip}, {recipes}, and {workflows}

``` r
citation("parsnip")
#>
#> To cite package 'parsnip' in publications use:
#>
#>   Kuhn M, Vaughan D (2022). _parsnip: A Common API to Modeling and
#>   Analysis Functions_. R package version 1.0.0,
#>   <https://CRAN.R-project.org/package=parsnip>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {parsnip: A Common API to Modeling and Analysis Functions},
#>     author = {Max Kuhn and Davis Vaughan},
#>     year = {2022},
#>     note = {R package version 1.0.0},
#>     url = {https://CRAN.R-project.org/package=parsnip},
#>   }
citation("recipes")
#>
#> To cite package 'recipes' in publications use:
#>
#>   Kuhn M, Wickham H (2022). _recipes: Preprocessing and Feature
#>   Engineering Steps for Modeling_. R package version 1.0.0,
#>   <https://CRAN.R-project.org/package=recipes>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {recipes: Preprocessing and Feature Engineering Steps for Modeling},
#>     author = {Max Kuhn and Hadley Wickham},
#>     year = {2022},
#>     note = {R package version 1.0.0},
#>     url = {https://CRAN.R-project.org/package=recipes},
#>   }
citation("workflows")
#>
#> To cite package 'workflows' in publications use:
#>
#>   Vaughan D (2022). _workflows: Modeling Workflows_. R package version
#>   1.0.0, <https://CRAN.R-project.org/package=workflows>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {workflows: Modeling Workflows},
#>     author = {Davis Vaughan},
#>     year = {2022},
#>     note = {R package version 1.0.0},
#>     url = {https://CRAN.R-project.org/package=workflows},
#>   }
```

- [ ] Line 50, add reference for {rsample}

``` r
citation("rsample")
#>
#> To cite package 'rsample' in publications use:
#>
#>   Silge J, Chow F, Kuhn M, Wickham H (2022). _rsample: General
#>   Resampling Infrastructure_. R package version 1.0.0,
#>   <https://CRAN.R-project.org/package=rsample>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {rsample: General Resampling Infrastructure},
#>     author = {Julia Silge and Fanny Chow and Max Kuhn and Hadley Wickham},
#>     year = {2022},
#>     note = {R package version 1.0.0},
#>     url = {https://CRAN.R-project.org/package=rsample},
#>   }
```

- [ ] Line 71, add reference to {ggplot2}

``` r
citation("ggplot2")
#>
#> To cite ggplot2 in publications, please use:
#>
#>   H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
#>   Springer-Verlag New York, 2016.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Book{,
#>     author = {Hadley Wickham},
#>     title = {ggplot2: Elegant Graphics for Data Analysis},
#>     publisher = {Springer-Verlag New York},
#>     year = {2016},
#>     isbn = {978-3-319-24277-4},
#>     url = {https://ggplot2.tidyverse.org},
#>   }
```

- [ ] Line 80, add reference to {h2o} (in addition to the current reference to LeDell & Poirier, 2020)

``` r
citation("h2o")
#>
#> To cite package 'h2o' in publications use:
#>
#>   LeDell E, Gill N, Aiello S, Fu A, Candel A, Click C, Kraljevic T,
#>   Nykodym T, Aboyoun P, Kurka M, Malohlava M (2022). _h2o: R Interface
#>   for the 'H2O' Scalable Machine Learning Platform_. R package version
#>   3.36.1.2, <https://CRAN.R-project.org/package=h2o>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {h2o: R Interface for the 'H2O' Scalable Machine Learning Platform},
#>     author = {Erin LeDell and Navdeep Gill and Spencer Aiello and Anqi Fu and Arno Candel and Cliff Click and Tom Kraljevic and Tomas Nykodym and Patrick Aboyoun and Michal Kurka and Michal Malohlava},
#>     year = {2022},
#>     note = {R package version 3.36.1.2},
#>     url = {https://CRAN.R-project.org/package=h2o},
#>   }
```

- [ ] Line 84, add reference to {mlr3}. {mlr3pipelines} seems properly cited already.

``` r
citation("mlr3")
#>
#> To cite package 'mlr3' in publications use:
#>
#>   Lang M, Binder M, Richter J, Schratz P, Pfisterer F, Coors S, Au Q,
#>   Casalicchio G, Kotthoff L, Bischl B (2019). "mlr3: A modern
#>   object-oriented machine learning framework in R." _Journal of Open
#>   Source Software_. doi:10.21105/joss.01903
#>   <https://doi.org/10.21105/joss.01903>,
#>   <https://joss.theoj.org/papers/10.21105/joss.01903>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Article{mlr3,
#>     title = {{mlr3}: A modern object-oriented machine learning framework in {R}},
#>     author = {Michel Lang and Martin Binder and Jakob Richter and Patrick Schratz and Florian Pfisterer and Stefan Coors and Quay Au and Giuseppe Casalicchio and Lars Kotthoff and Bernd Bischl},
#>     journal = {Journal of Open Source Software},
#>     year = {2019},
#>     month = {dec},
#>     doi = {10.21105/joss.01903},
#>     url = {https://joss.theoj.org/papers/10.21105/joss.01903},
#>   }
```
