### 2022-07-05 tidymodels/stacks#139 @osorensen
Further editiorial comments on JOSS submission
This is related to the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4471).

The reference list at the end seems incomplete:

- [ ] Line 111, Azure should be spelled with capital A. To fix this in BibTeX use an additional set of curly braces, i.e., replace `booktitle={Microsoft Azure Essentials},` with `booktitle={{Microsoft} {Azure} {Essentials}},`

https://github.com/tidymodels/stacks/blob/88be8e3223887b2f87aefd633462305473659eee/data-raw/paper/paper.bib#L101

Also add a URL to the book.

- [ ] Line 116-117, please include all authors in the citation, and also include a DOI. Click "Cite this article" at https://link.springer.com/article/10.1007/s11704-019-8208-z to get a working BibTeX citation.

- [ ] Lines 118-119, please include publisher the publisher, and preferrably also country.

- [ ] Lines 120-121, please include all author names. Here is the suggested citation from the JMLR website:

```
@article{10.5555/1953048.2078195,
author = {Pedregosa, Fabian and Varoquaux, Ga\"{e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, \'{E}douard},
title = {Scikit-Learn: Machine Learning in Python},
year = {2011},
issue_date = {2/1/2011},
publisher = {JMLR.org},
volume = {12},
number = {null},
issn = {1532-4435},
abstract = {Scikit-learn is a Python module integrating a wide range of state-of-the-art machine learning algorithms for medium-scale supervised and unsupervised problems. This package focuses on bringing machine learning to non-specialists using a general-purpose high-level language. Emphasis is put on ease of use, performance, documentation, and API consistency. It has minimal dependencies and is distributed under the simplified BSD license, encouraging its use in both academic and commercial settings. Source code, binaries, and documentation can be downloaded from http://scikit-learn.sourceforge.net.},
journal = {J. Mach. Learn. Res.},
month = {nov},
pages = {2825–2830},
numpages = {6}
}
```

- [ ] Lines 125-126. The DOI does not format as it should. Try replacing double quotes with curly braches here:

https://github.com/tidymodels/stacks/blob/88be8e3223887b2f87aefd633462305473659eee/data-raw/paper/paper.bib#L40
