+++ data-raw/paper/paper.bib
@@ -109,3 +109,80 @@ lastchecked = {2022-06-21}
   year={2015},
   howpublished={\url{https://keras.io}},
 }
+
+@Manual{parsnip,
+  title = {parsnip: A Common API to Modeling and Analysis Functions},
+  author = {Max Kuhn and Davis Vaughan},
+  year = {2022},
+  note = {R package version 1.0.0},
+  url = {https://CRAN.R-project.org/package=parsnip},
+}
+
+@Manual{recipes,
+  title = {recipes: Preprocessing and Feature Engineering Steps for Modeling},
+  author = {Max Kuhn and Hadley Wickham},
+  year = {2022},
+  note = {R package version 1.0.0},
+  url = {https://CRAN.R-project.org/package=recipes},
+}
+
+@Manual{workflows,
+  title = {workflows: Modeling Workflows},
+  author = {Davis Vaughan},
+  year = {2022},
+  note = {R package version 1.0.0},
+  url = {https://CRAN.R-project.org/package=workflows},
+}
+
+@Manual{rsample,
+  title = {rsample: General Resampling Infrastructure},
+  author = {Julia Silge and Fanny Chow and Max Kuhn and Hadley Wickham},
+  year = {2022},
+  note = {R package version 1.0.0},
+  url = {https://CRAN.R-project.org/package=rsample},
+}
+
+@Book{ggplot2,
+  author = {Hadley Wickham},
+  title = {ggplot2: Elegant Graphics for Data Analysis},
+  publisher = {Springer-Verlag New York},
+  year = {2016},
+  isbn = {978-3-319-24277-4},
+  url = {https://ggplot2.tidyverse.org},
+}
+
+@Manual{h2oR,
+  title = {h2o: R Interface for the 'H2O' Scalable Machine Learning Platform},
+  author = {Erin LeDell and Navdeep Gill and Spencer Aiello and Anqi Fu and Arno Candel and Cliff Click and Tom Kraljevic and Tomas Nykodym and Patrick Aboyoun and Michal Kurka and Michal Malohlava},
+  year = {2022},
+  note = {R package version 3.36.1.2},
+  url = {https://CRAN.R-project.org/package=h2o},
+}
+
+@Article{mlr3,
+  title = {{mlr3}: A modern object-oriented machine learning framework in {R}},
+  author = {Michel Lang and Martin Binder and Jakob Richter and Patrick Schratz and Florian Pfisterer and Stefan Coors and Quay Au and Giuseppe Casalicchio and Lars Kotthoff and Bernd Bischl},
+  journal = {Journal of Open Source Software},
+  year = {2019},
+  month = {dec},
+  doi = {10.21105/joss.01903},
+  url = {https://joss.theoj.org/papers/10.21105/joss.01903},
+}
+
+@Manual{tune,
+  title = {tune: Tidy Tuning Tools},
+  author = {Max Kuhn},
+  year = {2022},
+  note = {R package version 0.2.0},
+  url = {https://CRAN.R-project.org/package=tune},
+}
+
+
+
+
+
+
+
+
+
+
