+++ dev/pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.14.2
-pkgdown: 2.0.4
+pkgdown: 2.0.5
 pkgdown_sha: ~
 articles:
   basics: basics.html
   classification: classification.html
-last_built: 2022-06-22T15:25Z
+last_built: 2022-07-06T12:34Z
 urls:
   reference: https://stacks.tidymodels.org/reference
   article: https://stacks.tidymodels.org/articles
