+++ data-raw/paper/paper.bib
@@ -14,7 +14,7 @@
   number={43},
   pages={1686},
   year={2019},
-  doi="10.21105/joss.01686"
+  doi={10.21105/joss.01686}
 }
 
 @Manual{CRAN,
@@ -37,7 +37,7 @@
 
 @article{SuperLearner,
 author = {Mark J. van der Laan and Eric C Polley and Alan E. Hubbard},
-doi = "10.2202/1544-6115.1309",
+doi = {10.2202/1544-6115.1309},
 url = {https://doi.org/10.2202/1544-6115.1309},
 title = {Super {Learner}},
 journal = {Statistical Applications in Genetics and Molecular Biology},
@@ -46,7 +46,6 @@ volume = {6},
 year = {2007},
 lastchecked = {2022-06-21}
 }
-
 
 @Manual{chambers,
   title = {{John} {M.} {Chambers} {Statistical} {Software} {Award}},
@@ -61,7 +60,9 @@ lastchecked = {2022-06-21}
   journal={Monographs on statistics and applied probability},
   volume={143},
   pages={143},
-  year={2015}
+  year={2015},
+  publisher={{CRC} {Press}},
+  address={{Boca} {Raton}, {USA}}
 }
 
 @Article{mlr3ens,
@@ -76,17 +77,23 @@ lastchecked = {2022-06-21}
 }
 
 @article{jain2011scikit,
-  title={Scikit-learn: Machine learning in {Python}},
-  author={Jain, Arpan and others},
-  journal={J. Mach. Learn. Res},
-  volume={12},
-  pages={2825--2830},
-  year={2011}
+author = {Pedregosa, Fabian and Varoquaux, Ga\"{e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, \'{E}douard},
+title = {Scikit-Learn: Machine Learning in Python},
+year = {2011},
+issue_date = {2/1/2011},
+publisher = {JMLR.org},
+volume = {12},
+number = {null},
+issn = {1532-4435},
+journal = {J. Mach. Learn. Res.},
+month = {nov},
+pages = {2825–2830},
+numpages = {6}
 }
 
 @article{dong2020survey,
   title={A survey on ensemble learning},
-  author={Dong, Xibin and others},
+  author={Dong, Xibin and Yu, Zhiwen and Cao, Wenming and Shi, Yifan and Ma, Qianli},
   journal={Frontiers of Computer Science},
   volume={14},
   number={2},
@@ -96,11 +103,12 @@ lastchecked = {2022-06-21}
 }
 
 @incollection{barnes2015azure,
-  title={Azure machine learning},
+  title={Azure {Machine} {Learning}},
   author={Barnes, Jeff},
-  booktitle={Microsoft Azure Essentials},
+  booktitle={{Microsoft} {Azure} {Essentials}},
   year={2015},
-  publisher={Microsoft}
+  publisher={Microsoft {Press}},
+  howpublished={{ISBN:} 978-0-7356-9817-8},
 }
 
 @misc{chollet2015keras,
@@ -176,13 +184,3 @@ lastchecked = {2022-06-21}
   note = {R package version 0.2.0},
   url = {https://CRAN.R-project.org/package=tune},
 }
-
-
-
-
-
-
-
-
-
-
