### 2022-07-06 tidymodels/stacks#138 @simonpcouch
Thanks for all of your comments! I've included all of your edits and will regenerate the PDF on the review issue in a moment.
