### 2022-07-06 tidymodels/stacks#139 @simonpcouch
Thanks for all of your comments! I've included all of your edits with one exception: the Barnes, 2015, _Azure Machine Learning_ book doesn't seem to have a canonical URL, so I've included its ISBN instead.

Will regenerate the PDF on the review issue in a moment!
