### 2022-07-06 #4471 @simonpcouch
Updating with changes in response to https://github.com/tidymodels/stacks/issues/138 and https://github.com/tidymodels/stacks/issues/139!
