### 2022-07-06 tidymodels/stacks#139 @osorensen
Excellent!
