### 2022-07-06 9e0fa06
Update paper.bib
Protection of lower case in package names, related to review https://github.com/openjournals/joss-reviews/issues/4471.
