+++ data-raw/paper/paper.bib
@@ -119,7 +119,7 @@ numpages = {6}
 }
 
 @Manual{parsnip,
-  title = {parsnip: A Common API to Modeling and Analysis Functions},
+  title = {{parsnip}: A Common API to Modeling and Analysis Functions},
   author = {Max Kuhn and Davis Vaughan},
   year = {2022},
   note = {R package version 1.0.0},
@@ -127,7 +127,7 @@ numpages = {6}
 }
 
 @Manual{recipes,
-  title = {recipes: Preprocessing and Feature Engineering Steps for Modeling},
+  title = {{recipes}: Preprocessing and Feature Engineering Steps for Modeling},
   author = {Max Kuhn and Hadley Wickham},
   year = {2022},
   note = {R package version 1.0.0},
@@ -135,7 +135,7 @@ numpages = {6}
 }
 
 @Manual{workflows,
-  title = {workflows: Modeling Workflows},
+  title = {{workflows}: Modeling Workflows},
   author = {Davis Vaughan},
   year = {2022},
   note = {R package version 1.0.0},
@@ -143,7 +143,7 @@ numpages = {6}
 }
 
 @Manual{rsample,
-  title = {rsample: General Resampling Infrastructure},
+  title = {{rsample}: General Resampling Infrastructure},
   author = {Julia Silge and Fanny Chow and Max Kuhn and Hadley Wickham},
   year = {2022},
   note = {R package version 1.0.0},
@@ -178,7 +178,7 @@ numpages = {6}
 }
 
 @Manual{tune,
-  title = {tune: Tidy Tuning Tools},
+  title = {{tune}: Tidy Tuning Tools},
   author = {Max Kuhn},
   year = {2022},
   note = {R package version 0.2.0},
