+++ dev/pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   basics: basics.html
   classification: classification.html
-last_built: 2022-07-06T13:15Z
+last_built: 2022-07-06T14:24Z
 urls:
   reference: https://stacks.tidymodels.org/reference
   article: https://stacks.tidymodels.org/articles
